+++ src/c/CudaMex/AddImageWith.cpp
@@ -3,11 +3,11 @@
 
 void AddImageWith::execute( int nlhs, mxArray* plhs[], int nrhs, const mxArray* prhs[] )
 {
-	Vec<unsigned int> imageDims1;
+	Vec<size_t> imageDims1;
 	ImageContainer* imageIn1, * imageOut;
 	HostPixelType* mexImageOut;
 	setupImagePointers(prhs[0],&imageIn1,&plhs[0],&mexImageOut,&imageOut);
-	Vec<unsigned int> imageDims2;
+	Vec<size_t> imageDims2;
 	ImageContainer* imageIn2;
 	setupImagePointers(prhs[1],&imageIn2);
